Add ContainerPersistorTrait and use it for Flowcell.
